Added enemy that triggers combat
Placed a red circle representing the enemy that partially blocks the path.

made the firstperson camera its own tag and gave it collision so it can collide with objects in 3d space

Gave the 3d red circle enemy a collision script that simply loads the combat scene. We can create a interface file for enemies in the future with stuff all enemy classes will need to help future development